String creation in RAM instead of on disk

+++ include/seq_suffix_array.h
@@ -3,12 +3,13 @@
 #ifndef SEQUENTIAL_SUFFIX_ARRAY_H
 #define SEQUENTIAL_SUFFIX_ARRAY_H
 
+#include <cstdint>
 #include <vector>
 
 /**
  * @brief Build the suffix array of a text in O(n log n) using Manber & Myers algorithm
  *
- * @param text  Input text as a vector of ints (each 0–255)
+ * @param text  Input text as a vector of bytes (each 0–255)
  * @param sa    Output: the suffix array (indices of sorted suffixes)
  * @param rank  Temporary workspace / inverse of sa during construction
  * @param cnt   Workspace for bucket counts
@@ -16,7 +17,7 @@
  * @param bh    Bucket head flags
  * @param b2h   Secondary bucket head flags
  */
-void build_suffix_array(const std::vector<int> &text, std::vector<int> &sa, std::vector<int> &rank,
+void build_suffix_array(const std::vector<uint8_t> &text, std::vector<int> &sa, std::vector<int> &rank,
                         std::vector<int> &cnt, std::vector<int> &next, std::vector<bool> &bh, std::vector<bool> &b2h);
 
 /**
@@ -27,6 +28,6 @@ void build_suffix_array(const std::vector<int> &text, std::vector<int> &sa, std:
  * @param rank  Workspace / inverse of sa
  * @param lcp   Output: lcp[i] = length of LCP(sa[i], sa[i-1])
  */
-void build_lcp(const std::vector<int> &text, const std::vector<int> &sa, std::vector<int> &rank, std::vector<int> &lcp);
+void build_lcp(const std::vector<uint8_t> &text, const std::vector<int> &sa, std::vector<int> &rank, std::vector<int> &lcp);
 
 #endif // SEQUENTIAL_SUFFIX_ARRAY_H
